lantiq: Make DM200 firmware partition larger.
This now covers three netgear configuration/log/data partitions,
increasing storage space by 640k.

The only disadvantage of this is that settings in the original firmware
image are now overwritten, but this should not be surprising. The
original firmware populates these partitions if they're empty, so
firmware downgrade still works.

The sysconfig partition could probably go too, but it's small.